Convert basic.sh e2e test to Go
